Changed:
-Main: Add two empty lines.

-Changed the class from "Direction" to "UserInput":
  1)Renamed the class;
  2)Add event Action<UserInput>? OnChangedDirection;
  3)Added the "UseKey" method with a piece of the logic of the "Run" method of the "Game" class;
  4)Made the "ChangeDirection" method private.

-Class "GameMap":Remove method "HeightBoard".

-Class "Game": Removed part of the logic of the "Run" method, which was transferred to the "UseKey" method of the "UserInput" class.
