+++ Snake/Game.cs
@@ -1,47 +1,40 @@
 ﻿using GameSnake.Components;
-using GameSnake.Extension;
 
 namespace GameSnake
 {
     public class Game
     {
+        private bool GameOver = false;
         private GameMap _gameMap;
-        private Direction _directory;
+        private UserInput _directory;
         private Score _score;
 
         public Game(int weight, int height)
         {
-            _directory = new Direction();
+            _directory = new UserInput();
             _gameMap = new GameMap(weight, height);
             _score = new Score(height);
+            _directory.OnChangedDirection += _gameMap.ChangeSnakeDirection;
             _gameMap.OnEatScore += _score.Increase;
         }
 
         public void Run()
         {
-            while (true)
+            while (!GameOver)
             {
-                if (Console.KeyAvailable)
-                {
-                    ConsoleKey key = Console.ReadKey().Key;
-                    var direct = key.ToDirection();
-                    if (_directory.ChangeDirection(direct))
-                    {
-                        _gameMap.ChangeSnakeDirection(_directory);
-                    }
-                }
-
                 _gameMap.Clear();
 
                 //Game over
                 if (_gameMap.GameOver)
                 {
+                    GameOver = true;
                     break;
                 }
 
+                _directory.UseKey();
                 _gameMap.Move();
-                _gameMap.Draw();
 
+                _gameMap.Draw();
                 _score.Draw();
 
                 Thread.Sleep(100);
@@ -53,7 +46,6 @@ namespace GameSnake
             _gameMap.DrawBoarder();
             _gameMap.Draw();
             _score.Draw();
-
         }
     }
 }
